fix(api): route MCP calls through cataloged tools
Agent-Logs-Url: https://github.com/gredice/gredice/sessions/ad2c8ed1-8a48-4d73-a057-db4071322c37
